refactor: merge tree and navigator tree
Separated navigator tree configuration into 2 modules
complicated any improvements. Here merged into single module.
